refactor for hapi 17, WIP
