+++ package.json
@@ -22,6 +22,7 @@
   "dependencies": {
     "ajv": "^5.3.0",
     "boom": "^7.1.1",
+    "bounce": "^1.2.0",
     "catbox-memory": "^3.1.1",
     "chai": "^4.1.2",
     "clone": "^2.1.1",
