+++ plugins/core/base/routes/health.js
@@ -1,0 +1,10 @@
+module.exports = {
+  path: '/health',
+  method: 'GET',
+  options: {
+    tags: ['api']
+  },
+  handler: (request, h) => {
+    return 'Q server is alive';
+  }
+}
