+++ plugins/core/base/routes/version.js
@@ -1,0 +1,9 @@
+const package = require('../../../../package.json');
+
+module.exports = {
+  path: '/version',
+  method: 'GET',
+  handler: (request, h) => {
+    return package.version;
+  }
+}
