--- routes/version.js
@@ -1,9 +1,0 @@
-const package = require('../package');
-
-module.exports = {
-  path: '/version',
-  method: 'GET',
-  handler: (request, reply) => {
-    reply(package.version);
-  }
-}
